+++ src/instructions/interrupts.rs
@@ -1,5 +1,8 @@
 //! Enabling and disabling interrupts
 
+#[cfg(feature = "inline_asm")]
+use core::arch::asm;
+
 /// Returns whether interrupts are enabled.
 #[inline]
 pub fn are_enabled() -> bool {
@@ -156,7 +159,12 @@ pub fn int3() {
 /// It can also cause memory/register corruption depending on the interrupt
 /// implementation (if it expects values/pointers to be passed in registers).
 #[cfg(feature = "inline_asm")]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "nightly", feature = "inline_asm"))))]
+#[cfg_attr(
+    feature = "doc_cfg",
+    doc(cfg(any(feature = "nightly", feature = "inline_asm")))
+)]
 pub unsafe fn software_interrupt<const NUM: u8>() {
-    asm!("int {num}", num = const NUM, options(nomem, nostack));
+    unsafe {
+        asm!("int {num}", num = const NUM, options(nomem, nostack));
+    }
 }