+++ src/instructions/segmentation.rs
@@ -1,63 +1,13 @@
 //! Provides functions to read and write segment registers.
 
-#[cfg(docsrs)]
-use crate::{
-    registers::control::Cr4Flags,
-    structures::gdt::{Descriptor, GlobalDescriptorTable},
-};
+pub use crate::registers::segmentation::{Segment, Segment64, CS, DS, ES, FS, GS, SS};
 use crate::{
     registers::model_specific::{FsBase, GsBase, Msr},
     structures::gdt::SegmentSelector,
     VirtAddr,
 };
-
-/// An x86 segment
-///
-/// Segment registers on x86 are 16-bit [`SegmentSelector`]s, which index into
-/// the [`GlobalDescriptorTable`]. The corresponding GDT entry is used to
-/// configure the segment itself. Note that most segmentation functionality is
-/// disabled in 64-bit mode. See the individual segments for more information.
-pub trait Segment {
-    /// Returns the current value of the segment register.
-    fn get_reg() -> SegmentSelector;
-    /// Reload the segment register. Depending on the segment, this may also
-    /// reconfigure the corresponding segment.
-    ///
-    /// ## Safety
-    ///
-    /// This function is unsafe because the caller must ensure that `sel`
-    /// is a valid segment descriptor, and that reconfiguring the segment will
-    /// not cause undefined behavior.
-    unsafe fn set_reg(sel: SegmentSelector);
-}
-
-/// An x86 segment which is actually used in 64-bit mode
-///
-/// While most segments are unused in 64-bit mode, the FS and GS segments are
-/// still partially used. Only the 64-bit segment base address is used, and this
-/// address can be set via the GDT, or by using the `FSGSBASE` instructions.
-pub trait Segment64: Segment {
-    /// MSR containing the segment base. This MSR can be used to set the base
-    /// when [`CR4.FSGSBASE`][Cr4Flags::FSGSBASE] is **not** set.
-    const BASE: Msr;
-    /// Reads the segment base address
-    ///
-    /// ## Exceptions
-    ///
-    /// If [`CR4.FSGSBASE`][Cr4Flags::FSGSBASE] is not set, this instruction will throw a `#UD`.
-    fn read_base() -> VirtAddr;
-    /// Writes the segment base address
-    ///
-    /// ## Exceptions
-    ///
-    /// If [`CR4.FSGSBASE`][Cr4Flags::FSGSBASE] is not set, this instruction will throw a `#UD`.
-    ///
-    /// ## Safety
-    ///
-    /// The caller must ensure that this write operation has no unsafe side
-    /// effects, as the segment base address might be in use.
-    unsafe fn write_base(base: VirtAddr);
-}
+#[cfg(feature = "inline_asm")]
+use core::arch::asm;
 
 macro_rules! get_reg_impl {
     ($name:literal, $asm_get:ident) => {
@@ -83,10 +33,14 @@ macro_rules! segment_impl {
 
             unsafe fn set_reg(sel: SegmentSelector) {
                 #[cfg(feature = "inline_asm")]
-                asm!(concat!("mov ", $name, ", {0:x}"), in(reg) sel.0, options(nostack, preserves_flags));
+                unsafe {
+                    asm!(concat!("mov ", $name, ", {0:x}"), in(reg) sel.0, options(nostack, preserves_flags));
+                }
 
                 #[cfg(not(feature = "inline_asm"))]
-                crate::asm::$asm_load(sel.0);
+                unsafe{
+                    crate::asm::$asm_load(sel.0);
+                }
             }
         }
     };
@@ -111,23 +65,19 @@ macro_rules! segment64_impl {
 
             unsafe fn write_base(base: VirtAddr) {
                 #[cfg(feature = "inline_asm")]
-                asm!(concat!("wr", $name, "base {}"), in(reg) base.as_u64(), options(nostack, preserves_flags));
+                unsafe{
+                    asm!(concat!("wr", $name, "base {}"), in(reg) base.as_u64(), options(nostack, preserves_flags));
+                }
 
                 #[cfg(not(feature = "inline_asm"))]
-                crate::asm::$asm_wr(base.as_u64());
+                unsafe{
+                    crate::asm::$asm_wr(base.as_u64());
+                }
             }
         }
     };
 }
 
-/// Code Segment
-///
-/// The segment base and limit are unused in 64-bit mode. Only the L (long), D
-/// (default operation size), and DPL (descriptor privilege-level) fields of the
-/// descriptor are recognized. So changing the segment register can be used to
-/// change privilege level or enable/disable long mode.
-#[derive(Debug)]
-pub struct CS;
 impl Segment for CS {
     get_reg_impl!("cs", x86_64_asm_get_cs);
 
@@ -141,66 +91,31 @@ impl Segment for CS {
     /// for 64-bit far calls/jumps in long-mode, AMD does not.
     unsafe fn set_reg(sel: SegmentSelector) {
         #[cfg(feature = "inline_asm")]
-        asm!(
-            "push {sel}",
-            "lea {tmp}, [1f + rip]",
-            "push {tmp}",
-            "retfq",
-            "1:",
-            sel = in(reg) u64::from(sel.0),
-            tmp = lateout(reg) _,
-            options(preserves_flags),
-        );
+        unsafe {
+            asm!(
+                "push {sel}",
+                "lea {tmp}, [1f + rip]",
+                "push {tmp}",
+                "retfq",
+                "1:",
+                sel = in(reg) u64::from(sel.0),
+                tmp = lateout(reg) _,
+                options(preserves_flags),
+            );
+        }
 
         #[cfg(not(feature = "inline_asm"))]
-        crate::asm::x86_64_asm_set_cs(u64::from(sel.0));
+        unsafe {
+            crate::asm::x86_64_asm_set_cs(u64::from(sel.0));
+        }
     }
 }
 
-/// Stack Segment
-///
-/// Entirely unused in 64-bit mode; setting the segment register does nothing.
-/// However, in ring 3, the SS register still has to point to a valid
-/// [`Descriptor`] (it cannot be zero). This means a user-mode read/write
-/// segment descriptor must be present in the GDT.
-///
-/// This register is also set by the `syscall`/`sysret` and
-/// `sysenter`/`sysexit` instructions (even on 64-bit transitions). This is to
-/// maintain symmetry with 32-bit transitions where setting SS actually will
-/// actually have an effect.
-#[derive(Debug)]
-pub struct SS;
 segment_impl!(SS, "ss", x86_64_asm_get_ss, x86_64_asm_load_ss);
-
-/// Data Segment
-///
-/// Entirely unused in 64-bit mode; setting the segment register does nothing.
-#[derive(Debug)]
-pub struct DS;
 segment_impl!(DS, "ds", x86_64_asm_get_ds, x86_64_asm_load_ds);
-
-/// ES Segment
-///
-/// Entirely unused in 64-bit mode; setting the segment register does nothing.
-#[derive(Debug)]
-pub struct ES;
 segment_impl!(ES, "es", x86_64_asm_get_es, x86_64_asm_load_es);
-
-/// FS Segment
-///
-/// Only base is used in 64-bit mode, see [`Segment64`]. This is often used in
-/// user-mode for Thread-Local Storage (TLS).
-#[derive(Debug)]
-pub struct FS;
 segment_impl!(FS, "fs", x86_64_asm_get_fs, x86_64_asm_load_fs);
 segment64_impl!(FS, "fs", FsBase, x86_64_asm_rdfsbase, x86_64_asm_wrfsbase);
-
-/// GS Segment
-///
-/// Only base is used in 64-bit mode, see [`Segment64`]. In kernel-mode, the GS
-/// base often points to a per-cpu kernel data structure.
-#[derive(Debug)]
-pub struct GS;
 segment_impl!(GS, "gs", x86_64_asm_get_gs, x86_64_asm_load_gs);
 segment64_impl!(GS, "gs", GsBase, x86_64_asm_rdgsbase, x86_64_asm_wrgsbase);
 
@@ -213,10 +128,14 @@ impl GS {
     /// swap operation cannot lead to undefined behavior.
     pub unsafe fn swap() {
         #[cfg(feature = "inline_asm")]
-        asm!("swapgs", options(nostack, preserves_flags));
+        unsafe {
+            asm!("swapgs", options(nostack, preserves_flags));
+        }
 
         #[cfg(not(feature = "inline_asm"))]
-        crate::asm::x86_64_asm_swapgs();
+        unsafe {
+            crate::asm::x86_64_asm_swapgs();
+        }
     }
 }
 
@@ -225,49 +144,49 @@ impl GS {
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn set_cs(sel: SegmentSelector) {
-    CS::set_reg(sel)
+    unsafe { CS::set_reg(sel) }
 }
 /// Alias for [`SS::set_reg()`]
 #[deprecated(since = "0.14.4", note = "use `SS::set_reg()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn load_ss(sel: SegmentSelector) {
-    SS::set_reg(sel)
+    unsafe { SS::set_reg(sel) }
 }
 /// Alias for [`DS::set_reg()`]
 #[deprecated(since = "0.14.4", note = "use `DS::set_reg()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn load_ds(sel: SegmentSelector) {
-    DS::set_reg(sel)
+    unsafe { DS::set_reg(sel) }
 }
 /// Alias for [`ES::set_reg()`]
 #[deprecated(since = "0.14.4", note = "use `ES::set_reg()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn load_es(sel: SegmentSelector) {
-    ES::set_reg(sel)
+    unsafe { ES::set_reg(sel) }
 }
 /// Alias for [`FS::set_reg()`]
 #[deprecated(since = "0.14.4", note = "use `FS::set_reg()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn load_fs(sel: SegmentSelector) {
-    FS::set_reg(sel)
+    unsafe { FS::set_reg(sel) }
 }
 /// Alias for [`GS::set_reg()`]
 #[deprecated(since = "0.14.4", note = "use `GS::set_reg()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn load_gs(sel: SegmentSelector) {
-    GS::set_reg(sel)
+    unsafe { GS::set_reg(sel) }
 }
 /// Alias for [`GS::swap()`]
 #[deprecated(since = "0.14.4", note = "use `GS::swap()` instead")]
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn swap_gs() {
-    GS::swap()
+    unsafe { GS::swap() }
 }
 /// Alias for [`CS::get_reg()`]
 #[deprecated(since = "0.14.4", note = "use `CS::get_reg()` instead")]
@@ -283,7 +202,7 @@ pub fn cs() -> SegmentSelector {
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn wrfsbase(val: u64) {
-    FS::write_base(VirtAddr::new(val))
+    unsafe { FS::write_base(VirtAddr::new(val)) }
 }
 /// Alias for [`FS::read_base()`]
 #[deprecated(since = "0.14.4", note = "use `FS::read_base()` instead")]
@@ -299,7 +218,7 @@ pub unsafe fn rdfsbase() -> u64 {
 #[allow(clippy::missing_safety_doc)]
 #[inline]
 pub unsafe fn wrgsbase(val: u64) {
-    GS::write_base(VirtAddr::new(val))
+    unsafe { GS::write_base(VirtAddr::new(val)) }
 }
 /// Alias for [`GS::read_base()`]
 #[deprecated(since = "0.14.4", note = "use `GS::read_base()` instead")]