+++ src/registers/model_specific.rs
@@ -1,12 +1,9 @@
 //! Functions to read and write model specific registers.
 
-#[cfg(docsrs)]
-use crate::{
-    instructions::segmentation::{Segment64, FS, GS},
-    registers::control::Cr4Flags,
-};
-
 use bitflags::bitflags;
+// imports for intra doc links
+#[cfg(doc)]
+use crate::registers::segmentation::{FS, GS};
 
 /// A model specific register.
 #[derive(Debug)]
@@ -30,13 +27,19 @@ pub struct FsBase;
 
 /// [GS].Base Model Specific Register.
 ///
-/// [`GS::swap`] swaps this register with [`KernelGsBase`].
+#[cfg_attr(
+    feature = "instructions",
+    doc = "[`GS::swap`] swaps this register with [`KernelGsBase`]."
+)]
 #[derive(Debug)]
 pub struct GsBase;
 
 /// KernelGsBase Model Specific Register.
 ///
-/// [`GS::swap`] swaps this register with [`GsBase`].
+#[cfg_attr(
+    feature = "instructions",
+    doc = "[`GS::swap`] swaps this register with [`GsBase`]."
+)]
 #[derive(Debug)]
 pub struct KernelGsBase;
 
@@ -119,6 +122,14 @@ mod x86_64 {
     use bit_field::BitField;
     use core::convert::TryInto;
     use core::fmt;
+    // imports for intra doc links
+    #[cfg(doc)]
+    use crate::registers::{
+        control::Cr4Flags,
+        segmentation::{Segment, Segment64, CS, SS},
+    };
+    #[cfg(feature = "inline_asm")]
+    use core::arch::asm;
 
     impl Msr {
         /// Read 64 bits msr register.
@@ -132,17 +143,21 @@ mod x86_64 {
             #[cfg(feature = "inline_asm")]
             {
                 let (high, low): (u32, u32);
-                asm!(
-                    "rdmsr",
-                    in("ecx") self.0,
-                    out("eax") low, out("edx") high,
-                    options(nomem, nostack, preserves_flags),
-                );
+                unsafe {
+                    asm!(
+                        "rdmsr",
+                        in("ecx") self.0,
+                        out("eax") low, out("edx") high,
+                        options(nomem, nostack, preserves_flags),
+                    );
+                }
                 ((high as u64) << 32) | (low as u64)
             }
 
             #[cfg(not(feature = "inline_asm"))]
-            crate::asm::x86_64_asm_rdmsr(self.0)
+            unsafe {
+                crate::asm::x86_64_asm_rdmsr(self.0)
+            }
         }
 
         /// Write 64 bits to msr register.
@@ -157,15 +172,19 @@ mod x86_64 {
             let high = (value >> 32) as u32;
 
             #[cfg(feature = "inline_asm")]
-            asm!(
-                "wrmsr",
-                in("ecx") self.0,
-                in("eax") low, in("edx") high,
-                options(nostack, preserves_flags),
-            );
+            unsafe {
+                asm!(
+                    "wrmsr",
+                    in("ecx") self.0,
+                    in("eax") low, in("edx") high,
+                    options(nostack, preserves_flags),
+                );
+            }
 
             #[cfg(not(feature = "inline_asm"))]
-            crate::asm::x86_64_asm_wrmsr(self.0, low, high);
+            unsafe {
+                crate::asm::x86_64_asm_wrmsr(self.0, low, high);
+            }
         }
     }
 
@@ -196,7 +215,9 @@ mod x86_64 {
             let reserved = old_value & !(EferFlags::all().bits());
             let new_value = reserved | flags.bits();
 
-            Self::write_raw(new_value);
+            unsafe {
+                Self::write_raw(new_value);
+            }
         }
 
         /// Write the EFER flags.
@@ -210,7 +231,9 @@ mod x86_64 {
         #[inline]
         pub unsafe fn write_raw(flags: u64) {
             let mut msr = Self::MSR;
-            msr.write(flags);
+            unsafe {
+                msr.write(flags);
+            }
         }
 
         /// Update EFER flags.
@@ -228,7 +251,9 @@ mod x86_64 {
         {
             let mut flags = Self::read();
             f(&mut flags);
-            Self::write(flags);
+            unsafe {
+                Self::write(flags);
+            }
         }
     }
 
@@ -353,7 +378,9 @@ mod x86_64 {
             msr_value.set_bits(48..64, sysret.into());
             msr_value.set_bits(32..48, syscall.into());
             let mut msr = Self::MSR;
-            msr.write(msr_value);
+            unsafe {
+                msr.write(msr_value);
+            }
         }
 
         /// Write the Ring 0 and Ring 3 segment bases.