+++ src/structures/paging/mapper/recursive_page_table.rs
@@ -4,12 +4,12 @@ use core::fmt;
 
 use super::*;
 use crate::registers::control::Cr3;
-use crate::structures::paging::PageTableIndex;
+use crate::structures::paging::page_table::PageTableLevel;
 use crate::structures::paging::{
     frame_alloc::FrameAllocator,
-    page::{AddressNotAligned, NotGiantPageSize},
+    page::{AddressNotAligned, NotGiantPageSize, PageRangeInclusive},
     page_table::{FrameError, PageTable, PageTableEntry, PageTableFlags},
-    Page, PageSize, PhysFrame, Size1GiB, Size2MiB, Size4KiB,
+    FrameDeallocator, Page, PageSize, PageTableIndex, PhysFrame, Size1GiB, Size2MiB, Size4KiB,
 };
 use crate::VirtAddr;
 
@@ -339,8 +339,6 @@ impl<'a> Mapper<Size1GiB> for RecursivePageTable<'a> {
         Ok((frame, MapperFlush::new(page)))
     }
 
-    // allow unused_unsafe until https://github.com/rust-lang/rfcs/pull/2585 lands
-    #[allow(unused_unsafe)]
     unsafe fn update_flags(
         &mut self,
         page: Page<Size1GiB>,
@@ -467,8 +465,6 @@ impl<'a> Mapper<Size2MiB> for RecursivePageTable<'a> {
         Ok((frame, MapperFlush::new(page)))
     }
 
-    // allow unused_unsafe until https://github.com/rust-lang/rfcs/pull/2585 lands
-    #[allow(unused_unsafe)]
     unsafe fn update_flags(
         &mut self,
         page: Page<Size2MiB>,
@@ -526,7 +522,7 @@ impl<'a> Mapper<Size2MiB> for RecursivePageTable<'a> {
             return Err(FlagUpdateError::PageNotMapped);
         }
 
-        let p3 = &mut *(p3_ptr(page, self.recursive_index));
+        let p3 = unsafe { &mut *(p3_ptr(page, self.recursive_index)) };
         let p3_entry = &mut p3[page.p3_index()];
 
         if p3_entry.is_unused() {
@@ -625,8 +621,6 @@ impl<'a> Mapper<Size4KiB> for RecursivePageTable<'a> {
         Ok((frame, MapperFlush::new(page)))
     }
 
-    // allow unused_unsafe until https://github.com/rust-lang/rfcs/pull/2585 lands
-    #[allow(unused_unsafe)]
     unsafe fn update_flags(
         &mut self,
         page: Page<Size4KiB>,
@@ -689,7 +683,7 @@ impl<'a> Mapper<Size4KiB> for RecursivePageTable<'a> {
             return Err(FlagUpdateError::PageNotMapped);
         }
 
-        let p3 = &mut *(p3_ptr(page, self.recursive_index));
+        let p3 = unsafe { &mut *(p3_ptr(page, self.recursive_index)) };
         let p3_entry = &mut p3[page.p3_index()];
 
         if p3_entry.is_unused() {
@@ -712,13 +706,13 @@ impl<'a> Mapper<Size4KiB> for RecursivePageTable<'a> {
             return Err(FlagUpdateError::PageNotMapped);
         }
 
-        let p3 = &mut *(p3_ptr(page, self.recursive_index));
+        let p3 = unsafe { &mut *(p3_ptr(page, self.recursive_index)) };
 
         if p3[page.p3_index()].is_unused() {
             return Err(FlagUpdateError::PageNotMapped);
         }
 
-        let p2 = &mut *(p2_ptr(page, self.recursive_index));
+        let p2 = unsafe { &mut *(p2_ptr(page, self.recursive_index)) };
         let p2_entry = &mut p2[page.p2_index()];
 
         if p2_entry.is_unused() {
@@ -831,6 +825,99 @@ impl<'a> Translate for RecursivePageTable<'a> {
             offset,
             flags,
         }
+    }
+}
+
+impl<'a> CleanUp for RecursivePageTable<'a> {
+    #[inline]
+    unsafe fn clean_up<D>(&mut self, frame_deallocator: &mut D)
+    where
+        D: FrameDeallocator<Size4KiB>,
+    {
+        unsafe {
+            self.clean_up_addr_range(
+                PageRangeInclusive {
+                    start: Page::from_start_address(VirtAddr::new(0)).unwrap(),
+                    end: Page::from_start_address(VirtAddr::new(0xffff_ffff_ffff_f000)).unwrap(),
+                },
+                frame_deallocator,
+            )
+        }
+    }
+
+    unsafe fn clean_up_addr_range<D>(
+        &mut self,
+        range: PageRangeInclusive,
+        frame_deallocator: &mut D,
+    ) where
+        D: FrameDeallocator<Size4KiB>,
+    {
+        fn clean_up(
+            recursive_index: PageTableIndex,
+            page_table: &mut PageTable,
+            level: PageTableLevel,
+            range: PageRangeInclusive,
+            frame_deallocator: &mut impl FrameDeallocator<Size4KiB>,
+        ) -> bool {
+            if range.is_empty() {
+                return false;
+            }
+
+            let table_addr = range
+                .start
+                .start_address()
+                .align_down(level.table_address_space_alignment());
+
+            let start = range.start.page_table_index(level);
+            let end = range.end.page_table_index(level);
+
+            if let Some(next_level) = level.next_lower_level() {
+                let offset_per_entry = level.entry_address_space_alignment();
+                for (i, entry) in page_table
+                    .iter_mut()
+                    .enumerate()
+                    .take(usize::from(end) + 1)
+                    .skip(usize::from(start))
+                    .filter(|(i, _)| {
+                        !(level == PageTableLevel::Four && *i == recursive_index.into())
+                    })
+                {
+                    if let Ok(frame) = entry.frame() {
+                        let start = table_addr + (offset_per_entry * (i as u64));
+                        let end = start + (offset_per_entry - 1);
+                        let start = Page::<Size4KiB>::containing_address(start);
+                        let start = start.max(range.start);
+                        let end = Page::<Size4KiB>::containing_address(end);
+                        let end = end.min(range.end);
+                        let page_table =
+                            [p1_ptr, p2_ptr, p3_ptr][level as usize - 2](start, recursive_index);
+                        let page_table = unsafe { &mut *page_table };
+                        if clean_up(
+                            recursive_index,
+                            page_table,
+                            next_level,
+                            Page::range_inclusive(start, end),
+                            frame_deallocator,
+                        ) {
+                            entry.set_unused();
+                            unsafe {
+                                frame_deallocator.deallocate_frame(frame);
+                            }
+                        }
+                    }
+                }
+            }
+
+            page_table.iter().all(PageTableEntry::is_unused)
+        }
+
+        clean_up(
+            self.recursive_index,
+            self.level_4_table_mut(),
+            PageTableLevel::Four,
+            range,
+            frame_deallocator,
+        );
     }
 }