+++ src/structures/paging/page_table.rs
@@ -268,7 +268,7 @@ impl fmt::Debug for PageTable {
 /// Can be used to select one of the 512 entries of a page table.
 ///
 /// Guaranteed to only ever contain 0..512.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct PageTableIndex(u16);
 
 impl PageTableIndex {
@@ -319,7 +319,7 @@ impl From<PageTableIndex> for usize {
 /// This type is returned by the `VirtAddr::page_offset` method.
 ///
 /// Guaranteed to only ever contain 0..4096.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct PageOffset(u16);
 
 impl PageOffset {
@@ -362,5 +362,40 @@ impl From<PageOffset> for usize {
     #[inline]
     fn from(offset: PageOffset) -> Self {
         usize::from(offset.0)
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+/// A value between 1 and 4.
+pub enum PageTableLevel {
+    /// Represents the level for a page table.
+    One = 1,
+    /// Represents the level for a page directory.
+    Two,
+    /// Represents the level for a page-directory pointer.
+    Three,
+    /// Represents the level for a page-map level-4.
+    Four,
+}
+
+impl PageTableLevel {
+    /// Returns the next lower level or `None` for level 1
+    pub const fn next_lower_level(self) -> Option<Self> {
+        match self {
+            PageTableLevel::Four => Some(PageTableLevel::Three),
+            PageTableLevel::Three => Some(PageTableLevel::Two),
+            PageTableLevel::Two => Some(PageTableLevel::One),
+            PageTableLevel::One => None,
+        }
+    }
+
+    /// Returns the alignment for the address space described by a table of this level.
+    pub const fn table_address_space_alignment(self) -> u64 {
+        1u64 << (self as u8 * 9 + 12)
+    }
+
+    /// Returns the alignment for the address space described by an entry in a table of this level.
+    pub const fn entry_address_space_alignment(self) -> u64 {
+        1u64 << (((self as u8 - 1) * 9) + 12)
     }
 }