+++ Changelog.md
@@ -1,5 +1,38 @@
 # Unreleased
 
+# 0.14.7 – 2021-12-18
+
+- fix: build error on the latest nightly ([#329](https://github.com/rust-osdev/x86_64/pull/329))
+- add `set_general_handler` macro ([#285](https://github.com/rust-osdev/x86_64/pull/285))
+- Derive common traits for number, range and enum types ([#315](https://github.com/rust-osdev/x86_64/pull/315))
+- Add the VMM Communication Exception (`#VC`) to the `InterruptDescriptorTable` ([#313](https://github.com/rust-osdev/x86_64/pull/313))
+- fix: enable manipulation of `InterruptStackFrame` ([#312](https://github.com/rust-osdev/x86_64/pull/312))
+- fix docs for `page_table_index` ([#318](https://github.com/rust-osdev/x86_64/pull/318))
+- Remove redundant alignment check ([#314](https://github.com/rust-osdev/x86_64/pull/314))
+- fix(idt): fix panic messages for `index` and `#VC` ([#321](https://github.com/rust-osdev/x86_64/pull/321))
+- remove `const_assert!` in favor of std's `assert!` ([#326](https://github.com/rust-osdev/x86_64/pull/326))
+- Move bootloader integration test to separate CI job ([#330](https://github.com/rust-osdev/x86_64/pull/330))
+
+# 0.14.6 – 2021-09-20
+
+- New `registers::segmentation` module ([#309](https://github.com/rust-osdev/x86_64/pull/309)), containing:
+  - `instructions::segmentation::{Segment, Segment64, CS, DS, ES, FS, GS, SS}`
+  - `structures::gdt::SegmentSelector`
+  - Old locations still re-export all the types, so this is not a breaking change.
+- Fixes build so that `cargo doc --no-default-features` succeeds.
+
+# 0.14.5 – 2021-09-04
+
+- Add `ExceptionVector` enum and additional flags to `PageFaultErrorCode` ([#303](https://github.com/rust-osdev/x86_64/pull/303))
+- Add `clean_up` and `clean_up_with_filter` methods to deallocate unused page tables ([#264](https://github.com/rust-osdev/x86_64/pull/264))
+- Rename some XCr0 and CR4 flags (#[275](https://github.com/rust-osdev/x86_64/pull/275))
+- Expose `MapperFlush::new` and `MapperFlushAll::new` constructor functions ([#296](https://github.com/rust-osdev/x86_64/pull/296))
+- Use `#[cfg(doc)]` instead of docs.rs-specific cfg flag (#[287](https://github.com/rust-osdev/x86_64/pull/287))
+- Some documentation updates:
+  - Update segment register references in `GDT::load*` method to non-deprecated methods ([#301](https://github.com/rust-osdev/x86_64/pull/301))
+  - Remove a panic note ([#300](https://github.com/rust-osdev/x86_64/pull/300))
+- Update `bit_field` dependency ([#306](https://github.com/rust-osdev/x86_64/pull/306))
+
 # 0.14.4 – 2021-07-19
 
 - Add `instructions::tables::sgdt` ([#279](https://github.com/rust-osdev/x86_64/pull/279))